# apps/api/docs/swagger.yaml
basePath: /
definitions:
  handler.HealthResponse:
    properties:
      service:
        type: string
      status:
        type: string
      uptime:
        type: string
    type: object
  handler.authResponse:
    properties:
      token:
        type: string
    type: object
  handler.callbackRequest:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
  handler.castQAVoteRequest:
    properties:
      audience_uid:
        type: string
      value:
        type: integer
    type: object
  handler.castQAVoteResponse:
    properties:
      action:
        type: string
      id:
        type: string
      qa_entry_id:
        type: string
      vote_value:
        type: integer
      voter_uid:
        type: string
    type: object
  handler.castVoteRequest:
    properties:
      audience_uid:
        type: string
      option_ids:
        items:
          type: string
        type: array
    type: object
  handler.castVoteResponse:
    properties:
      message:
        type: string
    type: object
  handler.createPollOption:
    properties:
      label:
        type: string
      position:
        type: integer
    type: object
  handler.createPollRequest:
    properties:
      answer_mode:
        type: string
      options:
        items:
          $ref: '#/definitions/handler.createPollOption'
        type: array
      question:
        type: string
      time_limit_sec:
        type: integer
    type: object
  handler.createQARequest:
    properties:
      body:
        type: string
      entry_type:
        type: string
    type: object
  handler.createQAResponse:
    properties:
      author_uid:
        type: string
      body:
        type: string
      created_at:
        type: string
      entry_type:
        type: string
      id:
        type: string
      is_hidden:
        type: boolean
      score:
        type: integer
      session_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  handler.createSessionRequest:
    properties:
      title:
        type: string
    type: object
  handler.createSessionResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  handler.joinSessionResponse:
    properties:
      audience_uid:
        type: string
      session_title:
        type: string
    type: object
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.moderateQARequest:
    properties:
      is_hidden:
        type: boolean
      status:
        type: string
    type: object
  handler.pollOptionResponse:
    properties:
      id:
        type: string
      label:
        type: string
      position:
        type: integer
      vote_count:
        type: integer
    type: object
  handler.pollResponse:
    properties:
      answer_mode:
        type: string
      created_at:
        type: string
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/handler.pollOptionResponse'
        type: array
      question:
        type: string
      session_id:
        type: string
      status:
        type: string
      time_limit_sec:
        type: integer
      updated_at:
        type: string
    type: object
  handler.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  handler.updatePollRequest:
    properties:
      status:
        type: string
    type: object
  models.Poll:
    properties:
      answer_mode:
        type: string
      created_at:
        type: string
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/models.PollOption'
        type: array
      question:
        type: string
      session_id:
        type: string
      status:
        type: string
      time_limit_sec:
        type: integer
      updated_at:
        type: string
    type: object
  models.PollOption:
    properties:
      id:
        type: string
      label:
        type: string
      poll_id:
        type: string
      position:
        type: integer
    type: object
  models.Session:
    properties:
      closed_at:
        type: string
      code:
        type: string
      created_at:
        type: string
      host:
        $ref: '#/definitions/models.User'
      host_id:
        type: string
      id:
        type: string
      polls:
        items:
          $ref: '#/definitions/models.Poll'
        type: array
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@livepulse.local
    name: API Support
    url: http://localhost:3000/support
  description: Interactive polling platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:3000/terms/
  title: LivePulse API
  version: "1.0"
paths:
  /healthz:
    get:
      consumes:
      - application/json
      description: Returns the health status of the API service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HealthResponse'
      summary: Health check endpoint
      tags:
      - health
  /v1/auth/callback:
    post:
      consumes:
      - application/json
      description: Find or create user from OAuth provider and return JWT
      parameters:
      - description: OAuth callback data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.callbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: OAuth callback
      tags:
      - auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password to get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - auth
  /v1/sessions:
    get:
      consumes:
      - application/json
      description: Get all sessions created by the authenticated host
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Session'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List sessions for host
      tags:
      - sessions
    post:
      consumes:
      - application/json
      description: Create a new polling session for the authenticated host
      parameters:
      - description: Session creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.createSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new session
      tags:
      - sessions
  /v1/sessions/{code}:
    get:
      consumes:
      - application/json
      description: Get a session by its unique code (public access)
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Session'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get session by code
      tags:
      - sessions
  /v1/sessions/{code}/join:
    post:
      consumes:
      - application/json
      description: Join a session as an audience member and get an audience UID
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Client ID for session tracking
        in: header
        name: X-Client-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.joinSessionResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Join a session
      tags:
      - sessions
  /v1/sessions/{code}/polls:
    get:
      consumes:
      - application/json
      description: List all polls in a session (optional auth to detect host)
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.pollResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List polls in a session
      tags:
      - polls
    post:
      consumes:
      - application/json
      description: Create a new poll in a session (host only)
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Poll creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createPollRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.pollResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new poll
      tags:
      - polls
  /v1/sessions/{code}/polls/{pollID}:
    get:
      consumes:
      - application/json
      description: Get a poll with its options and current vote counts
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Poll ID
        in: path
        name: pollID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.pollResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a poll by ID
      tags:
      - polls
    patch:
      consumes:
      - application/json
      description: Update poll status (activate, close, etc.) - host only
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Poll ID
        in: path
        name: pollID
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updatePollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.pollResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update poll status
      tags:
      - polls
  /v1/sessions/{code}/polls/{pollID}/vote:
    post:
      consumes:
      - application/json
      description: Submit a vote on an active poll (audience only, requires valid
        audience UID)
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Poll ID
        in: path
        name: pollID
        required: true
        type: string
      - description: Vote data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.castVoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.castVoteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cast a vote on a poll
      tags:
      - votes
  /v1/sessions/{code}/qa:
    get:
      consumes:
      - application/json
      description: List active Q&A entries for a session with cursor pagination
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List Q&A entries
      tags:
      - qa
    post:
      consumes:
      - application/json
      description: Submit a question or comment to a session
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Q&A entry data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createQARequest'
      - description: Audience UID
        in: header
        name: X-Audience-UID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.createQAResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a Q&A entry
      tags:
      - qa
  /v1/sessions/{code}/qa/{id}:
    patch:
      consumes:
      - application/json
      description: Host can pin, answer, hide, or unhide a Q&A entry
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Moderation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.moderateQARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.createQAResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Moderate a Q&A entry
      tags:
      - qa
  /v1/sessions/{code}/qa/{id}/vote:
    post:
      consumes:
      - application/json
      description: Cast an upvote or downvote on a question (toggle behavior)
      parameters:
      - description: Session code
        in: path
        name: code
        required: true
        type: string
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Vote data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.castQAVoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.castQAVoteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Vote on a Q&A entry
      tags:
      - qa
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"